Use vault-action to get gpg private_key and passphrase (#18)
I am not sure that it's the good way to do that because of certificates
(stored in Lenstra repo) needed to connect the GitHub action to Vault. I
copied certificates from lenstra repo into this one, but what it that
does if the certificates change? So, I need to think a little more...

see https://github.com/Lenstra/lenstra/pull/1371